requirements/requirements.txt

albumentations
#captum
colorama
dask[all]
einops
fastai
# ffcv
kornia
lightning-bolts
pytorch_lightning
attrs
icecream
jpeg4py
# =======
# -e git+https://github.com/JacobARose/solo-learn.git@ffcv

hydra-core==1.1.1
matplotlib
modin
more_itertools
munch
omegaconf
opencv-python
pandarallel
pandas
Pillow
rich
# =======
antlr4-python3-runtime
iniconfig
plotly
pluggy
plumbum
prettytable
py
pytest
python-dotenv
PyYAML
flake8
pre-commit
ray
Sphinx
scikit-learn
scikit-image
seaborn
stuf
# tensorflow>=2
timm
toml
toolz
torchinfo
torchmetrics
torch-summary

# Corresponds to torchdatasets==0.2.0 except includes minor fix to allow torch<1.9
# # git+git://github.com/szymonmaszke/torchdatasets.git@0389feea2c23657440f9ba9a1af976f436aa3440
swifter
tqdm
wandb
webdataset

